chore: move events_list.dart to \lists subdirectory
