+++ my_flutter/lib/src/resources/event_api_provider.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'package:http/http.dart' show Client;
-import 'package:my_flutter/src/models/events_list.dart';
+import 'package:my_flutter/src/models/lists/events_list.dart';
 import 'dart:convert';
 import 'util/ServerUrls.dart';
